requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
openai>=1.0.0
python-pptx>=0.6.21
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
pydantic>=2.0.0
python-dotenv>=1.0.0
